+++ CHANGELOG.md
@@ -120,3 +120,4 @@
 - 2023-10-05: Add deals, gp, scheduler, session, +3 cfg
 - 2023-10-06: Log chat, gp, qualifier, scheduler, +3
 - 2023-10-06: Add qualifier, scheduler, gp, deals, +2 cfg
+- 2023-10-07: Pagination gp, session, deals, scheduler, +3